requirements.txt
Django>=4.2
djangorestframework>=3.14
Pillow>=10.0
celery>=5.3
redis>=5.0
djangorestframework-simplejwt>=5.3
